+++ packages/security/src/immutability.ts
@@ -1,6 +1,7 @@
 import type { Context } from '@aeriajs/types'
 import type { SecurityCheckProps, SecurityCheckReadPayload, SecurityCheckWritePayload } from './types.js'
-import { ACErrors, ObjectId } from '@aeriajs/types'
+import { ACErrors } from '@aeriajs/types'
+import { ObjectId } from '@aeriajs/api'
 import { left, right, isLeft } from '@aeriajs/common'
 
 const internalCheckImmutability = async (