+++ packages/types/src/index.ts
@@ -11,9 +11,6 @@ export * from './property.js'
 export * from './schema.js'
 export * from './security.js'
 export * from './validation.js'
-export {
-  ObjectId,
-} from 'mongodb'
 export type {
   WithId, OptionalId,
 } from 'mongodb'