+++ packages/aeria-build/CHANGELOG.md
@@ -1,5 +1,26 @@
 # aeria-build
 
+## 0.0.13
+
+### Patch Changes
+
+- Remove unused deps from dependency tree
+- Updated dependencies
+  - @aeriajs/builtins@0.0.10
+  - @aeriajs/common@0.0.4
+  - @aeriajs/types@0.0.3
+  - @aeriajs/api@0.0.10
+
+## 0.0.12
+
+### Patch Changes
+
+- Updated dependencies
+  - @aeriajs/types@0.0.2
+  - @aeriajs/api@0.0.9
+  - @aeriajs/builtins@0.0.9
+  - @aeriajs/common@0.0.3
+
 ## 0.0.11
 
 ### Patch Changes