+++ packages/http/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @aeriajs/http
 
+## 0.0.7
+
+### Patch Changes
+
+- Remove unused deps from dependency tree
+- Updated dependencies
+  - @aeriajs/access-control@0.0.4
+  - @aeriajs/validation@0.0.7
+  - @aeriajs/common@0.0.4
+  - @aeriajs/types@0.0.3
+
+## 0.0.6
+
+### Patch Changes
+
+- Updated dependencies
+  - @aeriajs/validation@0.0.6
+  - @aeriajs/types@0.0.2
+  - @aeriajs/access-control@0.0.3
+  - @aeriajs/common@0.0.3
+
 ## 0.0.5
 
 ### Patch Changes