+++ packages/node-http/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @aeriajs/node-http
 
+## 0.0.7
+
+### Patch Changes
+
+- Remove unused deps from dependency tree
+- Updated dependencies
+  - @aeriajs/http@0.0.7
+
+## 0.0.6
+
+### Patch Changes
+
+- @aeriajs/http@0.0.6
+
 ## 0.0.5
 
 ### Patch Changes