+++ packages/server/CHANGELOG.md
@@ -1,5 +1,34 @@
 # @aeriajs/server
 
+## 0.0.10
+
+### Patch Changes
+
+- Remove unused deps from dependency tree
+- Updated dependencies
+  - @aeriajs/access-control@0.0.4
+  - @aeriajs/entrypoint@0.0.4
+  - @aeriajs/builtins@0.0.10
+  - @aeriajs/common@0.0.4
+  - @aeriajs/server@0.0.10
+  - @aeriajs/types@0.0.3
+  - @aeriajs/http@0.0.7
+  - @aeriajs/api@0.0.10
+
+## 0.0.9
+
+### Patch Changes
+
+- Updated dependencies
+  - @aeriajs/types@0.0.2
+  - @aeriajs/api@0.0.9
+  - @aeriajs/builtins@0.0.9
+  - @aeriajs/http@0.0.6
+  - @aeriajs/access-control@0.0.3
+  - @aeriajs/common@0.0.3
+  - @aeriajs/entrypoint@0.0.3
+  - @aeriajs/server@0.0.9
+
 ## 0.0.8
 
 ### Patch Changes