+++ packages/types/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @aeriajs/types
 
+## 0.0.3
+
+### Patch Changes
+
+- Remove unused deps from dependency tree
+
+## 0.0.2
+
+### Patch Changes
+
+- Drop uneeded dependencies
+
 ## 0.0.1
 
 ### Patch Changes