+++ packages/validation/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @aeriajs/validation
 
+## 0.0.7
+
+### Patch Changes
+
+- Remove unused deps from dependency tree
+- Updated dependencies
+  - @aeriajs/common@0.0.4
+  - @aeriajs/types@0.0.3
+
+## 0.0.6
+
+### Patch Changes
+
+- Drop uneeded dependencies
+- Updated dependencies
+  - @aeriajs/types@0.0.2
+  - @aeriajs/common@0.0.3
+
 ## 0.0.5
 
 ### Patch Changes